requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.12.0
pypdfium2==4.25.0
pytesseract==0.3.10
Pillow>=10.1.0,<10.2.0
groq==0.4.2
httpx==0.27.0
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==4.1.0
chromadb==0.4.22
python-multipart==0.0.6
aiofiles>=22.0,<24.0
python-dotenv==1.0.0